Fix flaky tab writing behavior using text/tabwriter (#25)
* fix flaky tab writing behavior

This fixes problems with tab writing caused by using a regular os.File writer. By standardizing the tab writer’s usage and ensuring consistent initialization, we improve the reliability and consistency of output formatting a lot.

* view: improve tabwriter documentation

* view: add the ability to debug the tabwriter

In the past, we’ve faced too many issues with flaky tab formatting. If the `debug` flag is specified, we should enable debugging for the tabwriter as well, leveraging its built-in support for this.

* view: move newline handling back to renderer

* view: preserve comment